requirements.txt
streamlit==1.31.0
opencv-python==4.9.0.80
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.2.0
cmake==3.28.1
dlib==19.24.2